doc string changes, and removing smoke test.
